+++ app/components/atom/Logo.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Logo() {
+  return <h1 className="text-2xl font-bold">SES勤怠管理アプリ（仮）</h1>;
+}
+
+export default Logo;
